feat: add support for showing app icons
This took quite a bit of back and forth with Claude Sonnet 4.5, and
manual experimentation/troubleshooting, to arrive at a solution I'm
reasonably happy with. It's not great, and the deeply nested loops are
unsightly, but there's no better alternative given the edge cases this
needs to handle in order to determine the icon in most scenarios. I'm
pretty sure it doesn't handle every scenario (e.g. AppImage, Flatpak, Snaps,
etc.), but there's an escape hatch by allowing the user to manually
place an icon file in a specific location (e.g.
`~/.local/share/icons/hicolor/scalable/apps`).

It is loosely based on [1].

[1]: https://github.com/LawnGnome/niri-taskbar/blob/874ed92a1711422bcaaf635c7c3316edfc6a9d31/src/icon.rs