feat: expose focused window directly
Previously only the currently focused window title and app ID were
exposed to QML, but this was limiting, since e.g. it wasn't easily
possible to look up the icon of the currently focused window.

So instead of adding a bunch more proxy methods, it's simpler if we
just expose the Window object directly.

This might need to be optimized, or expose null pointer bugs, but it
works well from my tests.